Replace requirejs-rails with webpacker
**Why**: requirejs-rails is not very well maintained, and is preventing this app from being upgraded to Rails 5. Webpacker is the new recommended way of handling JS.

**How**:
- Replace `requirejs-rails` gem with `webpacker`
- Run `bundle install` , then `bundle exec rake webpacker:install` and `bundle exec rake webpacker:install:erb`.
- Delete all unused JS in `vendor/assets/javascripts/polyfills/remy`
- Replace the RequireJS `routes` path with packs named after the controller and action

Note that this commit only renames the old JS files. The changes will come in a separate commit. Doing the renaming and changes in separate commits will allow git to see that the changes were made to the same file as opposed to treating it as a deletion of the old file and creation of the new file.